Устранение замечаний к ДЗ 2

+++ habitmanager/src/main/java/ru/list/Repository/DBImplementation/LogBookRepositoryDBImplementation.java
@@ -7,34 +7,42 @@ import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
 
+import ru.list.Db.DBConnection;
 import ru.list.Model.Habit;
 import ru.list.Model.LogBook;
 import ru.list.Model.Period;
 import ru.list.Model.Person;
 import ru.list.Repository.LogBookRepository;
+import ru.list.logger.Logger;
 
 public class LogBookRepositoryDBImplementation implements LogBookRepository {
-    private String nameTable = "logbook";
-    private String nameSchema = "habit";
-    private Connection connection = null;
-    private String nameSerialID = "logbook_id_seq";
+    private static final String nameTable = "habit.logbook";
+    private static final String habitTable = "habit.habit";
+    private static final String personTable = "habit.person";
+    private static final String nameSerialID = "habit.logbook_id_seq";
+    private DBConnection dbConnection = null;
+    private Logger logger = null;
 
-    public LogBookRepositoryDBImplementation(Connection connection) {
-        this.connection = connection;
-        if (nameSchema.length() > 0) {
-            nameTable = nameSchema + "." + nameTable;
-            nameSerialID = String.format("%s.%s", nameSchema, nameSerialID); 
-        }
+    public LogBookRepositoryDBImplementation(DBConnection dbConnection, Logger logger) {
+        this.dbConnection = dbConnection;
+        this.logger = logger;
     }
 
 
     @Override
     public boolean save(LogBook logBook) {
-        boolean result;
+        boolean result = false;
+        Connection connection = null;
+        PreparedStatement statement = null;
+        if (!dbConnection.connect()) {
+            logger.addRecord("Ошибка подключения к базе при добавлении записи в журнал", result);
+            return result;
+        }
         String sql = String.format("INSERT INTO %s (id, habit_id, date) VALUES (nextval('%s'), ?, ?)", nameTable,nameSerialID);
         try {
+            connection = dbConnection.getConnection();
             connection.setAutoCommit(false);
-            PreparedStatement statement = connection.prepareStatement(sql);
+            statement = connection.prepareStatement(sql);
             statement.setInt(1, logBook.getHabit().getId());
             statement.setDate(2, java.sql.Date.valueOf(logBook.getDate()));
             int count = statement.executeUpdate();
@@ -46,37 +54,72 @@ public class LogBookRepositoryDBImplementation implements LogBookRepository {
             try {
                 connection.rollback();
             } catch(SQLException ex) {
-                System.out.println("Ошибка подключения к БД: " + ex.getMessage());
+                logger.addRecord("Ошибка rollback: " + ex.getMessage(), result);
             }
+            logger.addRecord(" Ошибка выполенения запроса: " + e.getMessage(), result);
             result = false;
-        } 
+        } finally {
+            try {
+                statement.close();
+            } catch (SQLException e) {
+                logger.addRecord(" Ошибка закрытия PreparedStatement: " + e.getMessage(), result);
+            }
+        }
 
+        dbConnection.closeConnection();
         return result;
     }
 
     @Override
     public boolean delete(LogBook logBook) {
-        boolean result;
+        boolean result = false;
+        Connection connection = null;
+        PreparedStatement statement = null;
+        if (!dbConnection.connect()) {
+            logger.addRecord("Ошибка подключения к базе при удалении записи в журнал", result);
+            return result;
+        }
         String sql = "DELETE FROM " + nameTable + " WHERE id = ?";
         try {
-            PreparedStatement statement = connection.prepareStatement(sql);
+            connection = dbConnection.getConnection();
+            connection.setAutoCommit(false);
+            statement = connection.prepareStatement(sql);
             statement.setInt(1, logBook.getId());
             int count = statement.executeUpdate();
             result = (count == 1);
         } catch(SQLException e) {
+            try {
+                connection.rollback();
+            } catch(SQLException ex) {
+                logger.addRecord("Ошибка rollback: " + ex.getMessage(), result);
+            }
+            logger.addRecord(" Ошибка выполенения запроса: " + e.getMessage(), result);
             result = false;
+        } finally {
+            try {
+                statement.close();
+            } catch (SQLException e) {
+                logger.addRecord(" Ошибка закрытия PreparedStatement: " + e.getMessage(), result);
+            }
         }
 
+        dbConnection.closeConnection();
         return result;
     }
 
     @Override
     public List<LogBook> findByPerson(Person person) {
         List<LogBook> logBooks = new ArrayList<>();
-        String habitTable = "habit.habit";
+        Connection connection = null;
+        PreparedStatement statement = null;
+        if (!dbConnection.connect()) {
+            logger.addRecord("Ошибка подключения к базе при поиске записей в журнале", false);
+            return logBooks;
+        }
         String sql = String.format("SELECT * FROM %s l JOIN %s h ON l.habit_id = h.id WHERE h.person_id = ? ",nameTable,habitTable);
         try {
-            PreparedStatement statement = connection.prepareStatement(sql);
+            connection = dbConnection.getConnection();
+            statement = connection.prepareStatement(sql);
             statement.setInt(1, person.getId());
             ResultSet resultSet = statement.executeQuery();
             while (resultSet.next()) {
@@ -89,20 +132,33 @@ public class LogBookRepositoryDBImplementation implements LogBookRepository {
                 LogBook logBook = new LogBook(resultSet.getInt("id"), resultSet.getDate("date").toLocalDate(), habit);
                 logBooks.add(logBook);
             }
+            resultSet.close();
         } catch (SQLException e) {
-            System.out.println("Ошибка подключения к БД: " + e.getMessage());
+            logger.addRecord("Ошибка получения списка привычек: " + e.getMessage(),false);
+        } finally {
+            try {
+                statement.close();
+            } catch (SQLException e) {
+                logger.addRecord(" Ошибка закрытия PreparedStatement: " + e.getMessage(), false);
+            }
         }
+        dbConnection.closeConnection();
         return logBooks;
     }
 
     @Override
     public List<LogBook> findAll() {
         List<LogBook> logBooks = new ArrayList<>();
-        String habitTable = "habit.habit";
-        String personTable = "habit.person";
+        Connection connection = null;
+        PreparedStatement statement = null;
+        if (!dbConnection.connect()) {
+            logger.addRecord("Ошибка подключения к базе при поиске записей в журнале", false);
+            return logBooks;
+        }
         String sql = String.format("SELECT * FROM %s l JOIN %s h ON l.habit_id = h.id JOIN %s p ON h.person_id = p.id",nameTable,habitTable,personTable);
         try {
-            PreparedStatement statement = connection.prepareStatement(sql);
+            connection = dbConnection.getConnection();
+            statement = connection.prepareStatement(sql);
             ResultSet resultSet = statement.executeQuery();
             while (resultSet.next()) {
                 Person person = new Person(resultSet.getInt("person_id"),
@@ -120,18 +176,33 @@ public class LogBookRepositoryDBImplementation implements LogBookRepository {
                 LogBook logBook = new LogBook(resultSet.getInt("id"), resultSet.getDate("date").toLocalDate(), habit);
                 logBooks.add(logBook);
             }
+            resultSet.close();
         } catch (SQLException e) {
-            System.out.println("Ошибка подключения к БД: " + e.getMessage());
+            logger.addRecord("Ошибка получения списка привычек: " + e.getMessage(),false);
+        } finally {
+            try {
+                statement.close();
+            } catch (SQLException e) {
+                logger.addRecord(" Ошибка закрытия PreparedStatement: " + e.getMessage(), false);
+            }
         }
+        dbConnection.closeConnection();
         return logBooks;
     }
 
     @Override
     public List<LogBook> findByHabit(Habit habit) {
         List<LogBook> logBooks = new ArrayList<>();
+        Connection connection = null;
+        PreparedStatement statement = null;
+        if (!dbConnection.connect()) {
+            logger.addRecord("Ошибка подключения к базе при поиске записей в журнале", false);
+            return logBooks;
+        }
         String sql = String.format("SELECT * FROM %s WHERE habit_id = ? ",nameTable);
         try {
-            PreparedStatement statement = connection.prepareStatement(sql);
+            connection = dbConnection.getConnection();
+            statement = connection.prepareStatement(sql);
             statement.setInt(1, habit.getId());
             ResultSet resultSet = statement.executeQuery();
             while (resultSet.next()) {
@@ -139,8 +210,15 @@ public class LogBookRepositoryDBImplementation implements LogBookRepository {
                 logBooks.add(logBook);
             }
         } catch (SQLException e) {
-            System.out.println("Ошибка подключения к БД: " + e.getMessage());
+            logger.addRecord("Ошибка получения списка привычек: " + e.getMessage(),false);
+        } finally {
+            try {
+                statement.close();
+            } catch (SQLException e) {
+                logger.addRecord(" Ошибка закрытия PreparedStatement: " + e.getMessage(), false);
+            }
         }
+        dbConnection.closeConnection();
         return logBooks;
     }
