+++ habitmanager/src/main/java/ru/list/Repository/DBImplementation/PersonRepositoryDBImplementation.java
@@ -7,30 +7,36 @@ import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
 
+import ru.list.Db.DBConnection;
 import ru.list.Model.Person;
 import ru.list.Repository.PersonRepository;
+import ru.list.logger.Logger;
 
 public class PersonRepositoryDBImplementation implements PersonRepository {
-    private String nameTable = "person";
-    private String nameSchema = "habit";
-    private Connection connection = null;
-    private String nameSerialID = "person_id_seq";
+    private static final String nameTable = "habit.person";
+    private static final String nameSerialID = "habit.person_id_seq";
+    private DBConnection dbConnection = null;
+    private Logger logger = null;
 
-    public PersonRepositoryDBImplementation(Connection connection) {
-        this.connection = connection;
-        if (nameSchema.length() > 0) {
-            nameTable = nameSchema + "." + nameTable;
-            nameSerialID = String.format("%s.%s", nameSchema, nameSerialID); 
-        }
+    public PersonRepositoryDBImplementation(DBConnection dbConnection, Logger logger) {
+        this.dbConnection = dbConnection;
+        this.logger = logger;
     }
 
     @Override
     public boolean save(Person person) {
-        boolean result;
+        boolean result = false;
+        Connection connection = null;
+        PreparedStatement statement = null;
+        if (!dbConnection.connect()) {
+            logger.addRecord("Ошибка подключения к базе при добавлении пользователя", result);
+            return result;
+        }
         String sql = String.format("INSERT INTO %s (id, username, email, password, role, blocked) VALUES (nextval('%s'), ?, ?, ?, ?, ?)", nameTable,nameSerialID);
         try {
+            connection = dbConnection.getConnection();
             connection.setAutoCommit(false);
-            PreparedStatement statement = connection.prepareStatement(sql);
+            statement = connection.prepareStatement(sql);
             statement.setString(1, person.getName());
             statement.setString(2, person.getEmail());
             statement.setString(3, person.getPassword());
@@ -45,36 +51,75 @@ public class PersonRepositoryDBImplementation implements PersonRepository {
             try {
                 connection.rollback();
             } catch(SQLException ex) {
-                System.out.println("Ошибка подключения к БД: " + ex.getMessage());
+                logger.addRecord("Ошибка rollback: " + ex.getMessage(), result);
             }
+            logger.addRecord(" Ошибка выполенения запроса: " + e.getMessage(), result);
             result = false;
-        } 
+        } finally {
+            try {
+                statement.close();
+            } catch (SQLException e) {
+                logger.addRecord(" Ошибка закрытия PreparedStatement: " + e.getMessage(), result);
+            }
+        }
 
+        dbConnection.closeConnection();
         return result;
     }
 
     @Override
     public boolean delete(Person person) {
-        boolean result;
+        boolean result = false;
+        Connection connection = null;
+        PreparedStatement statement = null;
+        if (!dbConnection.connect()) {
+            logger.addRecord("Ошибка подключения к базе при удалении пользователя", result);
+            return result;
+        }
         String sql = "DELETE FROM " + nameTable + " WHERE id = ?";
         try {
-            PreparedStatement statement = connection.prepareStatement(sql);
+            connection = dbConnection.getConnection();
+            connection.setAutoCommit(false);
+            statement = connection.prepareStatement(sql);
             statement.setInt(1, person.getId());
             int count = statement.executeUpdate();
             result = (count == 1);
+            if (result) {
+                connection.commit();
+            }
         } catch(SQLException e) {
+            try {
+                connection.rollback();
+            } catch(SQLException ex) {
+                logger.addRecord("Ошибка rollback: " + ex.getMessage(), result);
+            }
+            logger.addRecord(" Ошибка выполенения запроса: " + e.getMessage(), result);
             result = false;
+        } finally {
+            try {
+                statement.close();
+            } catch (SQLException e) {
+                logger.addRecord(" Ошибка закрытия PreparedStatement: " + e.getMessage(), result);
+            }
         }
 
+        dbConnection.closeConnection();
         return result;
     }
 
     @Override
     public Person findByEmailAndPassword(String email, String password) {
         Person result = null;
+        Connection connection = null;
+        PreparedStatement statement = null;
+        if (!dbConnection.connect()) {
+            logger.addRecord("Ошибка подключения к базе при поиске пользователя", false);
+            return result;
+        }
         String sql = String.format("SELECT * FROM %s WHERE email = ? AND password = ?",nameTable);
         try {
-            PreparedStatement statement = connection.prepareStatement(sql);
+            connection = dbConnection.getConnection();
+            statement = connection.prepareStatement(sql);
             statement.setString(1, email);
             statement.setString(2, password);
             ResultSet resultSet = statement.executeQuery();
@@ -86,19 +131,34 @@ public class PersonRepositoryDBImplementation implements PersonRepository {
                                 resultSet.getInt("role"),
                                 resultSet.getBoolean("blocked"));
             }
-        } catch(SQLException e) {
-            result = null;
+            resultSet.close();
+        } catch (SQLException e) {
+            logger.addRecord("Ошибка получения пользователя: " + e.getMessage(),false);
+        } finally {
+            try {
+                statement.close();
+            } catch (SQLException e) {
+                logger.addRecord(" Ошибка закрытия PreparedStatement: " + e.getMessage(), false);
+            }
         }
 
+        dbConnection.closeConnection();
         return result;
     }
 
     @Override
     public Person findByPassword(String password) {
         Person result = null;
+        Connection connection = null;
+        PreparedStatement statement = null;
+        if (!dbConnection.connect()) {
+            logger.addRecord("Ошибка подключения к базе при поиске пользователя", false);
+            return result;
+        }
         String sql = String.format("SELECT * FROM %s WHERE password = ?",nameTable);
         try {
-            PreparedStatement statement = connection.prepareStatement(sql);
+            connection = dbConnection.getConnection();
+            statement = connection.prepareStatement(sql);
             statement.setString(1, password);
             ResultSet resultSet = statement.executeQuery();
             while (resultSet.next()) {
@@ -109,19 +169,34 @@ public class PersonRepositoryDBImplementation implements PersonRepository {
                                 resultSet.getInt("role"),
                                 resultSet.getBoolean("blocked"));
             }
+            resultSet.close();
         } catch(SQLException e) {
-            result = null;
+            logger.addRecord("Ошибка получения пользователя: " + e.getMessage(),false);
+        } finally {
+            try {
+                statement.close();
+            } catch (SQLException e) {
+                logger.addRecord(" Ошибка закрытия PreparedStatement: " + e.getMessage(), false);
+            }
         }
 
+        dbConnection.closeConnection();
         return result;
     }
 
     @Override
     public List<Person> findAll() {
         List<Person> result = new ArrayList<>();
+        Connection connection = null;
+        PreparedStatement statement = null;
+        if (!dbConnection.connect()) {
+            logger.addRecord("Ошибка подключения к базе при получении списка пользователей", false);
+            return result;
+        }
         String sql = "SELECT * FROM " + nameTable;
         try {
-            PreparedStatement statement = connection.prepareStatement(sql);
+            connection = dbConnection.getConnection();
+            statement = connection.prepareStatement(sql);
             ResultSet resultSet = statement.executeQuery();
             while (resultSet.next()) {
                 Person person = new Person(resultSet.getInt("id"),
@@ -132,28 +207,51 @@ public class PersonRepositoryDBImplementation implements PersonRepository {
                                 resultSet.getBoolean("blocked"));
                 result.add(person);
             }
+            resultSet.close();
         } catch(SQLException e) {
-            System.out.println("Ошибка подключения к БД: " + e.getMessage());
+            logger.addRecord("Ошибка получения списка пользователей: " + e.getMessage(),false);
+        } finally {
+            try {
+                statement.close();
+            } catch (SQLException e) {
+                logger.addRecord(" Ошибка закрытия PreparedStatement: " + e.getMessage(), false);
+            }
         }
 
+        dbConnection.closeConnection();
         return result;
     }
 
     @Override
     public boolean exist(Person person) {
         boolean result = false;
+        Connection connection = null;
+        PreparedStatement statement = null;
+        if (!dbConnection.connect()) {
+            logger.addRecord("Ошибка подключения к базе при поиске пользователя", false);
+            return result;
+        }
         String sql = String.format("SELECT COUNT(*) FROM %s WHERE id = ?",nameTable);
         try {
-            PreparedStatement statement = connection.prepareStatement(sql);
+            connection = dbConnection.getConnection();
+            statement = connection.prepareStatement(sql);
             statement.setInt(1, person.getId());
             ResultSet resultSet = statement.executeQuery();
             while (resultSet.next()) {
                 result = resultSet.getInt("COUNT") > 0;
             }
+            resultSet.close();
         } catch(SQLException e) {
-            result = false;
+            logger.addRecord("Ошибка поиска пользователя: " + e.getMessage(),false);
+        } finally {
+            try {
+                statement.close();
+            } catch (SQLException e) {
+                logger.addRecord(" Ошибка закрытия PreparedStatement: " + e.getMessage(), false);
+            }
         }
 
+        dbConnection.closeConnection();
         return result;
     }
